% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_bayesian_output.R
\name{BayesFactorTest}
\alias{BayesFactorTest}
\title{Fancy Table for two group comparison}
\usage{
BayesFactorTest(x, y, data)
}
\arguments{
\item{x}{two level variable}

\item{y}{variable to compare}

\item{data}{data.frame or tibble  used to create model fits. Used for capturing variable labels, if they exist}
}
\value{
A tibble with:  \code{Variable}, \code{Parameter}, \code{Median CI},
 \code{pd} (probability of direction), 
 \code{ROPE CI} (R.O.P.E and CI), 
 \code{ROPE\%} (R.O.P.E precentage), 
 \code{BF} (bayes factor),  
 \code{Prior \[loc scale\]} (Prior distribution with location and scale parameters)
}
\description{
Fancy Table for two group comparison
}
\details{
Two group comparison
}
\examples{

# Real World Examples
data(Bladder_Cancer)
library(bayestestR)
library(tidyverse)
Bladder_Cancer <- Bladder_Cancer \%>\% 
mutate(Cycles_cat = droplevels(Cycles_cat),
      Clinical_Stage_Model = recode_factor(Clinical_Stage_Grouped, 
                                           'Stage I/II (<=T2NxMx)' = 'Stage I/II (<=T2NxMx)',
                                           'Stage III (T3NxMx)' = 'Stage III/IV (T3/4NxMx)'
                                           ,'Stage IV (T4NxMx)' = 'Stage III/IV (T3/4NxMx)')
)

# For multiple variables use map_dfr function
vars_to_run <- c('PT0N0','Gender', 'Clinical_Stage_Model', 'Cycles_cat')

testmap <- purrr::map_dfr(
 vars_to_run, BayesFactorTest, y='Elix_Sum', data = Bladder_Cancer )

# kable(testmap, 'latex', booktabs = TRUE, linesep = '', 
# caption = 'Bayesian t-tests for Elix Sum')\%>\%
#  kable_styling(font_size = 7.0) 

# just one 
testone <- BayesFactorTest(x = 'Cycles_cat', y = 'Elix_Sum', data = Bladder_Cancer )

# kable(testone, 'latex', booktabs = TRUE, linesep = '', 
# caption = 'Bayesian t-tests for Elix Sum')\%>\%
# kable_styling(font_size = 7.0) 

}
