% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests_and_estimates.R
\name{cor_test}
\alias{cor_test}
\title{Correlation Test for Two Continuous Variables}
\usage{
cor_test(x, y, method = c("pearson", "kendall", "spearman"),
  seed = 68954857, B = 10000, exact = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{numeric vector (can include NA values)}

\item{y}{numeric vector (can include NA values)}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated}

\item{seed}{seed (only used if \code{method = "spearman"})}

\item{B}{number of reps (only used if \code{method = "spearman"})}

\item{exact}{if no ties should you do exact (TRUE) or asymptotic (FALSE) method. Exact method is the default (only used if \code{method = "spearman" (and no ties) or "kendall"})}

\item{verbose}{a logical variable indicating if warnings and messages should be displayed}
}
\value{
spearman_test pvalue
}
\description{
A wrapper for cor.test function, except if spearman selected and ties in at least one variable, in which case this is a wrapper for coin::spreaman_test in with approximate method.
}
\details{
To always get reproducible results when using approximate method we need to set seed inside of the call, and order the data
}
\examples{

set.seed(5432322)
x <- rnorm(20,0,3)
y <- x + rnorm(20,0,5)
cor_test(x,y, method = 'pearson')
cor_test(x,y, method = 'kendall')
cor_test(x,y, method = 'spearman')

}
