% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{DATA DICTIONARY
creates  print class and levels and whatnot}
\usage{
dictionary(dataset, NAcode = NA, printalllevels = FALSE)
}
\arguments{
\item{dataset}{A data frame name.}

\item{NAcode}{A quoted string ie "NA" or whatever your NA code is.}

\item{printalllevels}{TRUE OR FALSE print all levels of your variables}
}
\value{
names and column number of variables or matches.
}
\description{
DATA DICTIONARY
creates  print class and levels and whatnot
}
\examples{
data(Bladder_Cancer)
dictionary(Bladder_Cancer, "NA", printalllevels=FALSE)

}
