\name{exampleData_BAMA}
\alias{exampleData_BAMA}
\docType{data}
\title{
exampleData_BAMA
}
\description{
  A description of the dataset. 
}
\usage{data("exampleData_BAMA")}
\format{
  A data frame with 252 observations on the following 6 variables.
  \describe{
    \item{\code{pubID}}{a character vector}
    \item{\code{group}}{a numeric vector}
    \item{\code{visitno}}{a numeric vector}
    \item{\code{antigen}}{a character vector}
    \item{\code{magnitude}}{a numeric vector}
    \item{\code{response}}{a numeric vector}
  }
}
\details{
 more details than the description above 
}
\source{ Not sure
}
\references{ Don't know
}
\examples{
data(exampleData_BAMA)
## maybe str(exampleData_BAMA) ; plot(exampleData_BAMA) ...
}
\keyword{datasets}
