\name{exampleData_ICS}
\alias{exampleData_ICS}
\docType{data}
\title{
exampleData_ICS
}
\description{
  A description of the dataset. 
}
\usage{data("exampleData_ICS")}
\format{
  A data frame with 306 observations on the following 15 variables.
  \describe{
    \item{\code{pubID}}{a character vector}
    \item{\code{Group}}{a character vector}
    \item{\code{Visit}}{a numeric vector}
    \item{\code{Stim}}{a character vector}
    \item{\code{Parent}}{a character vector}
    \item{\code{Population}}{a character vector}
    \item{\code{Count}}{a numeric vector}
    \item{\code{ParentCount}}{a numeric vector}
    \item{\code{CountBG}}{a numeric vector}
    \item{\code{ParentCountBG}}{a numeric vector}
    \item{\code{PercentCell}}{a numeric vector}
    \item{\code{PercentCellNet}}{a numeric vector}
    \item{\code{response_prob}}{a numeric vector}
    \item{\code{response_fdr_P}}{a numeric vector}
    \item{\code{response}}{a numeric vector}
  }
}
\details{
 If necessary, more details than the  description above  
}
\source{ Not sure
}
\references{ Don't know
}
\examples{
data(exampleData_ICS)
## maybe str(exampleData_ICS) ; plot(exampleData_ICS) ...
}
\keyword{datasets}
