\name{exampleData_NAb}
\alias{exampleData_NAb}
\docType{data}
\title{
exampleData_NAb
}
\description{
 A concise  description of the dataset.  
}
\usage{data("exampleData_NAb")}
\format{
  A data frame with 210 observations on the following 9 variables.
  \describe{
    \item{\code{pubID}}{a character vector}
    \item{\code{group}}{a numeric vector}
    \item{\code{visitno}}{a numeric vector}
    \item{\code{celltype}}{a character vector}
    \item{\code{isolate}}{a character vector}
    \item{\code{titer_mod_50}}{a numeric vector}
    \item{\code{titer_mod_80}}{a numeric vector}
    \item{\code{response_50}}{a numeric vector}
    \item{\code{response_80}}{a numeric vector}
  }
}
\details{
 more details than the description  above  
}
\source{ Not sure
}
\references{ Don't know
}
\examples{
data(exampleData_NAb)
## maybe str(exampleData_NAb) ; plot(exampleData_NAb) ...
}
\keyword{datasets}
