% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibility_tables.R
\name{get_session_info}
\alias{get_session_info}
\title{Get Reproducibility Tables}
\usage{
get_session_info()
}
\value{
list of length two, containing dataframe of Software Session Information and dataframe of Software Package Version Information
}
\description{
Creating tables used at the end of reports, for reproducibility. Most of the information is based off of sessioninfo::session_info()
}
\details{
Both tables usually printing with \code{kable()} at the end of a report
}
\examples{

get_session_info()

}
