% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoffittFunctions.r
\docType{package}
\name{MoffittFunctions-package}
\alias{MoffittFunctions}
\alias{MoffittFunctions-package}
\title{MoffittFunctions: Moffitt Biostat Functions}
\description{
Statistical, data processing, and annotation functions for Moffitt Biostat.
}
\author{
\strong{Maintainer}: Jimmy Fulp \email{william.fulp@moffitt.org}

Authors:
\itemize{
  \item Ram Thapa \email{Ram.Thapa@moffitt.org}
}

}
\keyword{internal}
