% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentchange.R
\name{percentchange}
\alias{percentchange}
\title{Calculate percent change from baseline
testing adding a function to MoffittFunctions package}
\usage{
percentchange(baseline, nextone)
}
\arguments{
\item{baseline}{Baseline or starting value}

\item{nextone}{Next value. This could be a vector of values.}
}
\description{
Calculate percent change from baseline
testing adding a function to MoffittFunctions package
}
