% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mixed_models.R
\name{pretty_mixed_effects_coxme}
\alias{pretty_mixed_effects_coxme}
\title{Fancy Table Output of Mixed effects COX PH model}
\usage{
pretty_mixed_effects_coxme(fit, df)
}
\arguments{
\item{fit}{glmer fit}

\item{df}{is data.frame or tibble  used to create model fits. Used for capturing variable labels, if they exist}
}
\value{
A data.frame  with:    \code{Variable}, \code{Level}, \code{HR (95\% CI)}, \code{P Value} (for categorical variables comparing to reference), \code{Overall P Value} (for categorical variables with 3+ levels).
}
\description{
pretty_mixed_effects_coxme()  Cox mixed effects model using the coxme() function.  Hazard ratios are produced with confidence intervals. P values are also produced. For categorical variables with 3+ levels overall Type 3 p values are calculated, in addition to p values comparing to the first level (reference).
}
\examples{
# Mixed effect cox ph model example - random intercepts
library(magrittr)
library(ggeffects)
library(sjmisc)
library(lme4)
library(splines)
library(coxme)
library(car)
library(survival)
library(broom.mixed)
library(rlang)
library(kableExtra)
data(PDACdata)
set.seed(542542522)
#USE THIS ONE for 3 months to 2 years (24 months)
surv_months_obj <- survival::Surv(time = PDACdata$dx_lastcontact_death_months,
                                 event = PDACdata$puf_vital_status == 0)
   cactfit  <- coxme(surv_months_obj ~   
     caci3 + surgery_volume_recode +  (1|puf_facility_id),
     data = PDACdata,
                    control=coxme.control(eps = 1e-04, toler.chol = .Machine$double.eps^0.75,
                            iter.max = 200, inner.iter = Quote(max(4, fit0$iter+1)),
                            sparse.calc = NULL,
                            optpar = list(method = "BFGS", control=list(reltol = 1e-3)),
                            refine.df=4, refine.detail=FALSE, refine.method="control",
                            sparse=c(50, .02),
                            varinit=c(.02, .1, .4, .8)^2, corinit = c(0, .3))  )
 cactfitcmetable <-  pretty_mixed_effects_coxme(cactfit, PDACdata)
# for PDF output 
#  kable(cactfitcmetable , 'latex', escape = FALSE, longtable = T, booktabs = TRUE, linesep = '', 
#  caption = 'Full Mulitvariable Mixed effects Cox Proportional-Hazards Regression Model with 
#  interaction for Overall Survival2 to 10 years')

}
