% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_mixed_models.R
\name{pretty_mixed_effects_glmer}
\alias{pretty_mixed_effects_glmer}
\title{Fancy Table Output of Mixed effects model from glmer}
\usage{
pretty_mixed_effects_glmer(fit, df, rdintterm)
}
\arguments{
\item{fit}{glmer fit}

\item{df}{is data.frame or tibble  used to create model fits. Used for capturing variable labels, if they exist}

\item{rdintterm}{random term in the model}
}
\value{
A data.frame with: \code{Variable}, \code{Level}, \code{ OR (95\% CI)}, \code{P Value} (for categorical variables comparing to reference), \code{Overall P Value} (for categorical variables with 3+ levels).
}
\description{
pretty_mixed_effects_glmer() takes a mixed effects model from glmer() fit object and calculates estimates, odds ratios with confidence intervals. P values are also produced. For categorical variables with 3+ levels overall Type 3 p values are calculated, in addition to p values comparing to the first level (reference).
}
\examples{

# glmer fit ####
library(magrittr)
library(ggeffects)
library(sjmisc)
library(lme4)
library(splines)
library(coxme)
library(car) 
library(survival)
library(broom.mixed)
library(rlang)
library(kableExtra)
data(PDACdata)
gmlmerfit <- glmer(I(PDACdata$mortality90day)=="Dead" ~   caci3
                  + surgery_volume_recode
                  + (1|puf_facility_id)
                  ,family = binomial(link = "logit"), 
                  data = PDACdata )
 
# CALL function include fit name, data frame name and random intercept term name in quotes ####
outputtable <- pretty_mixed_effects_glmer(gmlmerfit, PDACdata, "puf_facility_id")

kableExtra::kable(outputtable, 'html', caption = 'My Table') \%>\% 
   kableExtra::collapse_rows(c(1:2), row_group_label_position = 'stack')

# PDF output
#kable(outputtable, 'latex', escape = FALSE, longtable = T, booktabs = TRUE, 
# linesep = '', caption = 'Multivariable Mixed Effects Logistic Regression 
# Model Results for 90 day mortality with CACI, Volume, and Facility type')\%>\%
# footnote(number = c('OR are odds of death within 90 days
#  after the most definitive primary site surgery'))


}
