% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_output_functions.R
\name{pretty_pvalues_compareGroups}
\alias{pretty_pvalues_compareGroups}
\title{pretty_pvalues() wrapper for compareGroups table}
\usage{
pretty_pvalues_compareGroups(compareGroups_table_obj, digits = 3,
  include_p = FALSE, trailing_zeros = TRUE, output_type = NULL,
  bold = FALSE, italic = FALSE, background = NULL,
  sig_alpha = 0.05)
}
\arguments{
\item{compareGroups_table_obj}{createTable object created from compareGroups::createTable() function, with at least one p value column}

\item{digits}{number of digits to round to; values with zeros past this number of digits are truncated}

\item{include_p}{TRUE or FALSE: set to TRUE to print "p = " before each p-value}

\item{trailing_zeros}{TRUE or FALSE: default = TRUE, p-values are formatted with trailing zeros to the defined number of digits (i.e. 0.100 instead of 0.1 if digits= 3)}

\item{output_type}{output type, either NULL (default), "latex", or "html" (making special charaters latex friendly)}

\item{bold}{TRUE or FALSE: set to TRUE to bold p-values < the defined significance level}

\item{italic}{TRUE or FALSE: set to TRUE to italicize p-values < the defined significance level}

\item{background}{highlight color for p-values < the defined significance level. Default = NULL (no highlighting)}

\item{sig_alpha}{the defined significance level. Default = 0.05}
}
\value{
createTable object with rounding, formating, and highlighting of any p value column.
}
\description{
pretty_pvalues_compareGroups() takes a createTable object created from compareGroups::createTable(), and applies pretty_pvalues() to all p value columns.
}
\details{
This function is design specifically for input from compareGroups tables. Use \code{pretty_pvalues} directly when working with normal data.frames and other objects.
}
\examples{

data("Bladder_Cancer")
cycles_formula <- as.formula(Cycles_cat ~ Age_At_Diagnosis + Gender + Elix_Sum)
cycles_compare <- compareGroups::compareGroups(cycles_formula, data = Bladder_Cancer)
cycles_table <- compareGroups::createTable(cycles_compare, digits.p = 4, show.p.mul = T)
cycles_table_fancy <- pretty_pvalues_compareGroups(cycles_table, background = 'yellow')

# Printing createTable object in report
compareGroups::export2latex(cycles_table_fancy, size = 'footnotesize', label = 'tab:Variable-of-Interest-Table', caption = 'Comparing Variables to Number of Cycles', header.labels = c('p.overall' = 'Overall P'), landscape = F)

}
