% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_output_functions.R
\name{run_pretty_km_output}
\alias{run_pretty_km_output}
\title{Wrapper for KM Model Output, with Log-Rank p value}
\usage{
run_pretty_km_output(strata_in = NA, model_data, time_in, event_in,
  event_level = NULL, time_est = NULL, group_name = NULL,
  title_name = NULL, conf_level = 0.95, surv_est_prefix = "Time",
  surv_est_digits = 2, median_est_digits = 1, p_digits = 4,
  output_type = NULL, sig_alpha = 0.05, background = "yellow", ...)
}
\arguments{
\item{strata_in}{name of strata variable, or NA (default) if no strata desired}

\item{model_data}{dataset that contains \code{strata_in}, \code{time_in}, and \code{event_in} variables}

\item{time_in}{name of time variable component of outcome measure}

\item{event_in}{name of event status variable. If \code{event_level} = NULL then this must be the name of a FALSE/TRUE or 0/1 variable, where FALSE or 0 are considered the censored level, respectively}

\item{event_level}{event level for event status variable.}

\item{time_est}{numerical vector of time estimates. If NULL (default) no time estimates are calculated}

\item{group_name}{strata variable name. If NULL and strata exists then using variable}

\item{title_name}{title to use}

\item{conf_level}{the confidence level required (default is 0.95).}

\item{surv_est_prefix}{prefix to use in survival estimate names. Default is Time (i.e. Time:5, Time:10,...)}

\item{surv_est_digits}{number of digits to round p values for survival estimates for specified times}

\item{median_est_digits}{number of digits to round p values for Median Survival Estimates}

\item{p_digits}{number of digits to round p values for Log-Rank p value}

\item{output_type}{output type, either NULL (default), "latex", or "html" (making special charaters latex friendly)}

\item{sig_alpha}{the defined significance level. Default = 0.05}

\item{background}{background color of significant values, or no highlighting if NULL. Default is "yellow"}

\item{...}{other params to pass to \code{pretty_pvalues} (i.e. \code{bold} or \code{italic})}
}
\value{
A tibble with: \code{Name} (if provided), \code{Group} (if strata variable in fit), \code{Level} (if strata variable in fit), \code{Time:X} (Survival estimates for each time provided), \code{Median Estimate}. In no strata variable tibble is one row, otherwise nrows = number of strata levels.
}
\description{
This function takes a dataset, along with variables names for time and event status for KM fit, and possibly strata
}
\examples{

# Basic survival model examples
set.seed(542542522)
ybin <- sample(0:1, 100, replace = TRUE)
ybin2 <- sample(0:1, 100, replace = TRUE)
ybin3 <- sample(c('Dead','Alive'), 100, replace = TRUE)
y <- rexp(100,.1)
x1 <- factor(sample(LETTERS[1:2],100,replace = TRUE))
x2 <- factor(sample(letters[1:4],100,replace = TRUE))
my_data <- data.frame(y, ybin, ybin2, ybin3, x1, x2)
Hmisc::label(my_data$x1) <- "X1 Variable"

 # Single runs 
run_pretty_km_output(strata_in = 'x1', model_data = my_data, 
     time_in = 'y', event_in = 'ybin', time_est = NULL)
run_pretty_km_output(strata_in = 'x1', model_data = my_data, 
     time_in = 'y', event_in = 'ybin', time_est = c(5,10))
run_pretty_km_output(strata_in = 'x2', model_data = my_data, 
     time_in = 'y', event_in = 'ybin3', event_level = 'Dead', time_est = c(5,10))

# Multiple runs for different variables
library(dplyr) 
vars_to_run = c(NA, 'x1', 'x2')
purrr::map_dfr(vars_to_run, run_pretty_km_output, model_data = my_data,
     time_in = 'y', event_in = 'ybin', event_level = '0', time_est = NULL) \%>\% 
   select(Group, Level, everything())
   
km_info <- purrr::map_dfr(vars_to_run, run_pretty_km_output, model_data = my_data, time_in = 'y', 
     event_in = 'ybin3', event_level = 'Dead', time_est = c(5,10), surv_est_prefix = 'Year', 
     title_name = 'Overall Survival') \%>\% 
   select(Group, Level, everything())
   
km_info2 <- purrr::map_dfr(vars_to_run, run_pretty_km_output, model_data = my_data, time_in = 'y', 
     event_in = 'ybin2', time_est = c(5,10), surv_est_prefix = 'Year', 
     title_name = 'Cancer Specific Survival') \%>\% 
   select(Group, Level, everything())

options(knitr.kable.NA = '')
kableExtra::kable(bind_rows(km_info, km_info2), escape = FALSE, longtable = FALSE, booktabs = TRUE, linesep = '', 
     caption = 'Survival Percentage Estimates at 5 and 10 Years') \%>\%
  kableExtra::collapse_rows(c(1:2), row_group_label_position = 'stack', headers_to_remove = 1:2) 


  # Real World Example
  data(Bladder_Cancer)
  
  vars_to_run = c(NA, 'Gender', 'Clinical_Stage_Grouped', 'PT0N0', 'Any_Downstaging')
  
  purrr::map_dfr(vars_to_run, run_pretty_km_output, model_data = Bladder_Cancer, 
       time_in = 'Survival_Months', event_in = 'Vital_Status', event_level = 'Dead', 
       time_est = c(24,60), surv_est_prefix = 'Month', p_digits=5) \%>\% 
   select(Group, Level, everything())
  
}
