\name{testData_BAMA}
\alias{testData_BAMA}
\docType{data}
\title{
testData_BAMA
}
\description{
 A concise   description of the dataset. 
}
\usage{data("testData_BAMA")}
\format{
  A data frame with 270 observations on the following 6 variables.
  \describe{
    \item{\code{pubID}}{a character vector}
    \item{\code{group}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{antigen}}{a character vector}
    \item{\code{visit}}{a factor with levels \code{1} \code{2}}
    \item{\code{magnitude}}{a numeric vector}
    \item{\code{response}}{a numeric vector}
  }
}
\details{
 If necessary, more details than the __description__ above 
}
\source{ Not sure.
}
\references{ Don't know
}
\examples{
data(testData_BAMA)
## maybe str(testData_BAMA) ; plot(testData_BAMA) ...
}
\keyword{datasets}
