\name{testData_ICS}
\alias{testData_ICS}
\docType{data}
\title{
testData_ICS
}
\description{
 A concise  description of the dataset.  
}
\usage{data("testData_ICS")}
\format{
  A data frame with 198 observations on the following 15 variables.
  \describe{
    \item{\code{pubID}}{a character vector}
    \item{\code{Group}}{a factor with levels \code{1} \code{2} \code{3}}
    \item{\code{Visit}}{a factor with levels \code{1} \code{2}}
    \item{\code{Stim}}{a character vector}
    \item{\code{Parent}}{a character vector}
    \item{\code{Population}}{a character vector}
    \item{\code{Count}}{a numeric vector}
    \item{\code{ParentCount}}{a numeric vector}
    \item{\code{CountBG}}{a numeric vector}
    \item{\code{ParentCountBG}}{a numeric vector}
    \item{\code{PercentCell}}{a numeric vector}
    \item{\code{PercentCellNet}}{a numeric vector}
    \item{\code{response_prob}}{a numeric vector}
    \item{\code{response_fdr_P}}{a numeric vector}
    \item{\code{response}}{a numeric vector}
  }
}
\details{ more details than the  description  above  
}
\source{ Not sure
}
\references{ Don't know
}
\examples{
data(testData_ICS)
## maybe str(testData_ICS) ; plot(testData_ICS) ...
}
\keyword{datasets}
