\name{testData_NAb}
\alias{testData_NAb}
\docType{data}
\title{
testData_NAb
}
\description{
  A concise  description of the dataset.  
}
\usage{data("testData_NAb")}
\format{
  A data frame with 144 observations on the following 7 variables.
  \describe{
    \item{\code{celltype}}{a character vector}
    \item{\code{pubID}}{a character vector}
    \item{\code{visit}}{a factor with levels \code{1} \code{2}}
    \item{\code{group}}{a numeric vector}
    \item{\code{isolate}}{a character vector}
    \item{\code{magnitude}}{a numeric vector}
    \item{\code{response}}{a numeric vector}
  }
}
\details{ more details than the description above 
}
\source{ Not sure
}
\references{ Don't know
}
\examples{
data(testData_NAb)
## maybe str(testData_NAb) ; plot(testData_NAb) ...
}
\keyword{datasets}
