% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoffittFunctions-package.r
\docType{package}
\name{MoffittFunctions-package}
\alias{MoffittFunctions}
\alias{MoffittFunctions-package}
\title{Moffitt Functions}
\description{
Statistical, data processing, and annotation functions for data analysis.
Many functions for producing fancey tables in reports
}
\author{
\strong{Maintainer}: Jimmy Fulp \email{William.Fulp@moffitt.org}

Authors:
\itemize{
  \item Ram Thapa \email{Ram.Thapa@moffitt.org}
  \item Zachary Thompson \email{Zachary.Thompson@moffitt.org}
}

}
