% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoffittFunctions-package.r
\docType{package}
\name{MoffittFunctions-package}
\alias{MoffittFunctions}
\alias{MoffittFunctions-package}
\title{Moffitt Functions}
\description{
Statistical, data processing, and annotation functions for data analysis.
Many functions for producing fancey tables in reports
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.moffitt.usf.edu:8000/ReproducibleResearch/MoffittFunctions}
  \item Report bugs at \url{https://gitlab.moffitt.usf.edu:8000/ReproducibleResearch/MoffittFunctions/-/issues}
}

}
\author{
\strong{Maintainer}: Zachary Thompson \email{Zachary.Thompson@moffitt.org}

Authors:
\itemize{
  \item Jimmy Fulp \email{William.Fulp@moffitt.org}
  \item Ram Thapa \email{Ram.Thapa@moffitt.org}
}

}
